Move buttons out of component and onto page
